Folder Structure changed, Calculator contianer added and some mobile responsive changes done
