+++ src/App.jsx
@@ -1,4 +1,3 @@
-import { useState } from "react";
 import { BrowserRouter as Router, Routes, Route } from "react-router-dom";
 import Navbar from "./components/Navbar";
 import Footer from "./components/Footer";
@@ -7,7 +6,6 @@ import { About, Home, Blog, Careers, Patner, Contact } from "./pages";
 import "./Complete.css";
 
 function App() {
-  const [count, setCount] = useState(0);
 
   return (
     <>
