+++ src/pages/home/MobileContainer/style.css
@@ -1,0 +1,34 @@
+
+/* mobile-container */
+.mobile-container{
+    width: 100vw;
+    display: flex;
+    align-items: center;
+    justify-content: center;
+  
+  }
+  
+  .mobile-container .portfolios{
+    width: 95%;
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    justify-content: center;
+    background-color: #171c48;
+  }
+  
+  .mobile-container .portfolios .card p{
+      font-family: "Inter-SemiBold", "Inter", sans-serif;
+      font-size: 60px;
+      font-weight: 600;
+      letter-spacing: -2.4px;
+      line-height: 72px;
+      text-align: center;
+      color: rgb(255, 255, 255);
+  
+  }
+  
+  .mobile-container .portfolios .card .img-card img{
+    width: 1000px;
+    height: fit-content;
+  }
