+++ src/pages/home/VedioContainer/style.css
@@ -1,0 +1,149 @@
+:root {
+    --speed: 12000 ;
+   
+  }
+  
+  
+  /* container-1 css */
+  .container-1 {
+      /* height: 80vh; */
+      width: 100vw;
+      display: flex;
+      align-items: center;
+      justify-content: center;
+  }
+  .container-1 .framer {
+      display: flex;
+      gap: 10px;
+      flex-wrap: wrap;
+      align-items: center;
+      justify-content: center;
+      justify-content: space-around;
+      /* padding: 0px 60px 0px 60px; */
+  }
+  .container-1 .main-cards {
+      display: flex;
+      flex-direction: column;
+      justify-content: center;
+      align-items: center;
+  }
+    
+    .container-1 .main-cards .card-sub {
+      display: flex;
+      flex-wrap: wrap;
+    }
+    
+    .container-1 .card {
+      line-height: 2px;
+      display: flex;
+      flex-direction: column;
+      justify-content: center;
+      align-items: center;
+    }
+    
+    .container-1 .card h2 {
+      display: flex;
+      flex-wrap: wrap;
+      color: white;
+      font-size: 44px;
+    }
+    
+    .card-sub h2 {
+      display: flex;
+      flex-wrap: wrap;
+      justify-content: center;
+      align-items: center;
+      color: white;
+      font-size: 44px;
+    }
+    .card-sub h2 span {
+      font-size: 44px;
+    }
+    .card-sub {
+      display: flex;
+      flex-wrap: wrap;
+      align-items: center;
+      justify-content: center;
+      height: 80px;
+    }
+    
+    .card-sub .p1 {
+      display: flex;
+      justify-content: center;
+      align-items: center;
+      font-size: 28px;
+      font-family: "Montserrat", sans-serif;
+      font-weight: 700;
+      width: auto;
+      margin-top: 55px;
+    }
+    
+    .card-sub .p2 {
+      font-family: "Montserrat", sans-serif;
+      font-weight: 700;
+      width: auto;
+      font-size: 23px;
+      color: white;
+      display: flex;
+      line-height: 30px;
+      align-items: center;
+      justify-content: center;
+      flex-wrap: wrap;
+    }
+    .card-sub .p2 span{
+      font-size: 31px;
+    }
+  
+    .container-1 .framer .card .scroll-text{
+      height: 180px;
+      background-color: transparent;
+      display: flex;
+      justify-content: center;
+      align-items: center;
+    }
+    .container-1 .framer .card .scroll-text p{
+      margin: 0px;
+      font-family: "Inter-SemiBold", "Inter", sans-serif;
+      font-size: 28px;
+      font-weight: 600;
+      line-height: 40px;
+      align-items: center;
+    }
+   
+    
+    .card .card-vedio {
+      cursor: pointer;
+      padding: 0px 0px;
+      height: 325px;
+      /* margin-top: 70px; */
+      width: auto;
+    }
+  
+  
+    .card .card-vedio video {
+      border-radius: 35px;
+      padding: 25px;
+      height: 380px;
+      width: 100%;
+    }
+   
+    .card .card-btn {
+      gap: 18px;
+      margin-top: 30px;
+      display: flex;
+      justify-content: space-around;
+      flex-wrap: wrap;
+  }
+  .card .card-btn button {
+      cursor: pointer;
+      border: none;
+      text-decoration: none;
+      background-color: transparent;
+  }
+  .card .card-btn button img {
+      width: 260px;
+      height: 73px;
+  }
+
+
+
